# api/http/groupbuy_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.27.0
// source: api/http/groupbuy.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGroupBuyHTTPGetAllActivities = "/api.http.GroupBuyHTTP/GetAllActivities"
const OperationGroupBuyHTTPGetAllDiscounts = "/api.http.GroupBuyHTTP/GetAllDiscounts"
const OperationGroupBuyHTTPGetDiscountByID = "/api.http.GroupBuyHTTP/GetDiscountByID"
const OperationGroupBuyHTTPMarketTrial = "/api.http.GroupBuyHTTP/MarketTrial"

type GroupBuyHTTPHTTPServer interface {
	// GetAllActivities GetAllActivities retrieves all group buy activities
	GetAllActivities(context.Context, *GetAllActivitiesRequest) (*GetAllActivitiesReply, error)
	// GetAllDiscounts GetAllDiscounts retrieves all group buy discounts
	GetAllDiscounts(context.Context, *GetAllDiscountsRequest) (*GetAllDiscountsReply, error)
	// GetDiscountByID GetDiscountByID retrieves a group buy discount by ID
	GetDiscountByID(context.Context, *GetDiscountByIDRequest) (*GetDiscountByIDReply, error)
	// MarketTrial MarketTrial handles market trial requests
	MarketTrial(context.Context, *MarketTrialRequest) (*MarketTrialReply, error)
}

func RegisterGroupBuyHTTPHTTPServer(s *http.Server, srv GroupBuyHTTPHTTPServer) {
	r := s.Route("/")
	r.GET("/groupbuy/activities", _GroupBuyHTTP_GetAllActivities0_HTTP_Handler(srv))
	r.GET("/groupbuy/discounts", _GroupBuyHTTP_GetAllDiscounts0_HTTP_Handler(srv))
	r.GET("/groupbuy/discount", _GroupBuyHTTP_GetDiscountByID0_HTTP_Handler(srv))
	r.GET("/market/trial", _GroupBuyHTTP_MarketTrial0_HTTP_Handler(srv))
}

func _GroupBuyHTTP_GetAllActivities0_HTTP_Handler(srv GroupBuyHTTPHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetAllActivitiesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGroupBuyHTTPGetAllActivities)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAllActivities(ctx, req.(*GetAllActivitiesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAllActivitiesReply)
		return ctx.Result(200, reply)
	}
}

func _GroupBuyHTTP_GetAllDiscounts0_HTTP_Handler(srv GroupBuyHTTPHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetAllDiscountsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGroupBuyHTTPGetAllDiscounts)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAllDiscounts(ctx, req.(*GetAllDiscountsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAllDiscountsReply)
		return ctx.Result(200, reply)
	}
}

func _GroupBuyHTTP_GetDiscountByID0_HTTP_Handler(srv GroupBuyHTTPHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetDiscountByIDRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGroupBuyHTTPGetDiscountByID)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetDiscountByID(ctx, req.(*GetDiscountByIDRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetDiscountByIDReply)
		return ctx.Result(200, reply)
	}
}

func _GroupBuyHTTP_MarketTrial0_HTTP_Handler(srv GroupBuyHTTPHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MarketTrialRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGroupBuyHTTPMarketTrial)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.MarketTrial(ctx, req.(*MarketTrialRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MarketTrialReply)
		return ctx.Result(200, reply)
	}
}

type GroupBuyHTTPHTTPClient interface {
	// GetAllActivities GetAllActivities retrieves all group buy activities
	GetAllActivities(ctx context.Context, req *GetAllActivitiesRequest, opts ...http.CallOption) (rsp *GetAllActivitiesReply, err error)
	// GetAllDiscounts GetAllDiscounts retrieves all group buy discounts
	GetAllDiscounts(ctx context.Context, req *GetAllDiscountsRequest, opts ...http.CallOption) (rsp *GetAllDiscountsReply, err error)
	// GetDiscountByID GetDiscountByID retrieves a group buy discount by ID
	GetDiscountByID(ctx context.Context, req *GetDiscountByIDRequest, opts ...http.CallOption) (rsp *GetDiscountByIDReply, err error)
	// MarketTrial MarketTrial handles market trial requests
	MarketTrial(ctx context.Context, req *MarketTrialRequest, opts ...http.CallOption) (rsp *MarketTrialReply, err error)
}

type GroupBuyHTTPHTTPClientImpl struct {
	cc *http.Client
}

func NewGroupBuyHTTPHTTPClient(client *http.Client) GroupBuyHTTPHTTPClient {
	return &GroupBuyHTTPHTTPClientImpl{client}
}

// GetAllActivities GetAllActivities retrieves all group buy activities
func (c *GroupBuyHTTPHTTPClientImpl) GetAllActivities(ctx context.Context, in *GetAllActivitiesRequest, opts ...http.CallOption) (*GetAllActivitiesReply, error) {
	var out GetAllActivitiesReply
	pattern := "/groupbuy/activities"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGroupBuyHTTPGetAllActivities))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetAllDiscounts GetAllDiscounts retrieves all group buy discounts
func (c *GroupBuyHTTPHTTPClientImpl) GetAllDiscounts(ctx context.Context, in *GetAllDiscountsRequest, opts ...http.CallOption) (*GetAllDiscountsReply, error) {
	var out GetAllDiscountsReply
	pattern := "/groupbuy/discounts"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGroupBuyHTTPGetAllDiscounts))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetDiscountByID GetDiscountByID retrieves a group buy discount by ID
func (c *GroupBuyHTTPHTTPClientImpl) GetDiscountByID(ctx context.Context, in *GetDiscountByIDRequest, opts ...http.CallOption) (*GetDiscountByIDReply, error) {
	var out GetDiscountByIDReply
	pattern := "/groupbuy/discount"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGroupBuyHTTPGetDiscountByID))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// MarketTrial MarketTrial handles market trial requests
func (c *GroupBuyHTTPHTTPClientImpl) MarketTrial(ctx context.Context, in *MarketTrialRequest, opts ...http.CallOption) (*MarketTrialReply, error) {
	var out MarketTrialReply
	pattern := "/market/trial"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGroupBuyHTTPMarketTrial))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}